+++ sdk/healthinsights/health-insights-radiologyinsights-rest/samples/v1/javascript/sample_complete_order_discrepancy_inference_async.js
@@ -137,7 +137,7 @@ function createRequestBody() {
 
   const patientDocumentData = {
     type: "note",
-    clinicalType: "radiologyReport",
+    clinicalType: ClinicalDocumentTypeEnum.RadiologyReport,
     id: "docid1",
     language: "en",
     authors: [authorData],