+++ sdk/healthinsights/health-insights-radiologyinsights-rest/test/public/SampleFindingInferenceAsync.spec.ts
@@ -4,7 +4,11 @@
 import { Recorder } from "@azure-tools/test-recorder";
 import { assert } from "chai";
 import { Context } from "mocha";
-import { AzureHealthInsightsClient, getLongRunningPoller } from "../../src";
+import {
+  AzureHealthInsightsClient,
+  ClinicalDocumentTypeEnum,
+  getLongRunningPoller,
+} from "../../src";
 import { createRecorder, createTestClient } from "./utils/recordedClient";
 
 const codingData = {
@@ -58,7 +62,7 @@ const content = {
 
 const patientDocumentData = {
   type: "note",
-  clinicalType: "radiologyReport",
+  clinicalType: ClinicalDocumentTypeEnum.RadiologyReport,
   id: "docid1",
   language: "en",
   authors: [authorData],